@@ -1,6 +1,6 @@
 package io.github.ianyong.spfdivisionalboundaries;
 
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.os.Bundle;
 
 import com.google.android.gms.maps.CameraUpdateFactory;
